requirements.txt
biopython
confuse
huggingface_hub
transformers
datasets
tokenizers
accelerate